+++ packages/varlet-eslint-config/index.js
@@ -49,7 +49,6 @@ module.exports = {
     'vue/no-v-html': 'off',
     'vue/attributes-order': 'off',
     'vue/require-default-prop': 'off',
-    'vue/no-unused-components': 'off',
     'vue/require-explicit-emits': 'off',
     '@typescript-eslint/camelcase': 'off',
     '@typescript-eslint/ban-ts-ignore': 'off',
@@ -60,5 +59,6 @@ module.exports = {
     '@typescript-eslint/no-non-null-assertion': 'off',
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@typescript-eslint/explicit-module-boundary-types': 'off',
+    '@typescript-eslint/no-empty-interface': 'off',
   },
 }