+++ packages/varlet-ui/src/tab-item/TabItem.vue
@@ -1,13 +1,12 @@
 <template>
-  <transition :name="transitionName" @after-enter="resetTransitionHeight">
-    <div class="var-tab-item" ref="tabItemEl" v-show="show">
-      <slot v-if="initSlot" />
-    </div>
-  </transition>
+  <var-swipe-item class="var-tab-item" :class="[!current && 'var-tab-item--inactive']">
+    <slot v-if="initSlot" />
+  </var-swipe-item>
 </template>
 
 <script lang="ts">
-import { defineComponent, ref, computed, watch } from 'vue'
+import SwipeItem from '../swipe-item'
+import { defineComponent, ref, computed } from 'vue'
 import { useTabsItems } from './provide'
 import { props } from './props'
 import type { Ref, ComputedRef } from 'vue'
@@ -15,51 +14,35 @@ import type { TabItemProvider } from './provide'
 
 export default defineComponent({
   name: 'VarTabItem',
+  components: {
+    [SwipeItem.name]: SwipeItem,
+  },
   props,
   setup(props) {
-    const show: Ref<boolean> = ref(false)
+    const current: Ref<boolean> = ref(false)
     const initSlot: Ref<boolean> = ref(false)
-    const tabItemEl: Ref<null | HTMLElement> = ref(null)
-    const transitionName: Ref<string> = ref('var-tabs-items-left')
     const name: ComputedRef<string | number | undefined> = computed(() => props.name)
-    const element: ComputedRef<HTMLElement | null> = computed(() => tabItemEl.value)
-    const { index, tabsItems, bindTabsItems } = useTabsItems()
-    const { active, resize, resetTransitionHeight } = tabsItems
+    const { index, bindTabsItems } = useTabsItems()
 
-    const transition = (newIndex: number, oldIndex: number) => {
-      transitionName.value =
-        oldIndex === -1 ? 'var-tab-item-right' : newIndex > oldIndex ? 'var-tab-item-right' : 'var-tab-item-left'
-      show.value = active.value === props.name || active.value === index.value
+    const setCurrent = (value: boolean) => {
+      if (!initSlot.value && value) {
+        initSlot.value = true
+      }
+
+      current.value = value
     }
 
     const tabItemProvider: TabItemProvider = {
       index,
       name,
-      element,
-      transition,
+      setCurrent,
     }
 
     bindTabsItems(tabItemProvider)
 
-    watch(() => props.name, resize)
-
-    watch(
-      () => show.value,
-      (newValue) => {
-        if (newValue) {
-          initSlot.value = true
-        }
-      },
-      { immediate: true }
-    )
-
     return {
-      tabItemEl,
+      current,
       initSlot,
-      show,
-      active,
-      transitionName,
-      resetTransitionHeight,
     }
   },
 })
@@ -67,5 +50,6 @@ export default defineComponent({
 
 <style lang="less">
 @import '../styles/common';
+@import '../swipe-item/swipeItem';
 @import './tabItem';
 </style>