+++ packages/varlet-ui/src/tabs-items/props.ts
@@ -5,6 +5,14 @@ export const props = {
     type: [String, Number],
     default: 0,
   },
+  canSwipe: {
+    type: Boolean,
+    default: true,
+  },
+  loop: {
+    type: Boolean,
+    default: false,
+  },
   'onUpdate:active': {
     type: Function as PropType<(active: string | number) => void>,
   },