+++ packages/varlet-ui/src/tabs-items/provide.ts
@@ -1,12 +1,7 @@
-import type { ComputedRef } from 'vue'
 import { useAtChildrenCounter, useChildren } from '../utils/components'
 import type { TabItemProvider } from '../tab-item/provide'
 
-export interface TabsItemsProvider {
-  active: ComputedRef<number | string>
-  resize(): void
-  resetTransitionHeight(): void
-}
+export interface TabsItemsProvider {}
 
 export const TABS_ITEMS_BIND_TAB_ITEM_KEY = Symbol('TABS_ITEMS_BIND_TAB_ITEM_KEY')
 export const TABS_ITEMS_COUNT_TAB_ITEM_KEY = Symbol('TABS_ITEMS_COUNT_TAB_ITEM_KEY')