+++ packages/varlet-ui/src/tabs/example/index.vue
@@ -95,20 +95,21 @@
     active-color="#fff"
     inactive-color="hsla(0, 0%, 100%, .6)"
     v-model:active="activeRelation"
+    offset-top="50"
   >
     <var-tab>{{ pack.option }}1</var-tab>
     <var-tab>{{ pack.option }}2</var-tab>
     <var-tab>{{ pack.option }}3</var-tab>
   </var-tabs>
 
-  <var-tabs-items v-model:active="activeRelation">
-    <var-tab-item style="padding: 8px 5px">
+  <var-tabs-items style="margin: 8px 0" v-model:active="activeRelation">
+    <var-tab-item style="padding: 0 4px">
       {{ pack.text }}
     </var-tab-item>
-    <var-tab-item style="padding: 8px 5px">
+    <var-tab-item style="padding: 0 4px">
       {{ pack.text2 }}
     </var-tab-item>
-    <var-tab-item style="padding: 8px 5px">
+    <var-tab-item style="padding: 0 4px">
       {{ pack.text3 }}
     </var-tab-item>
   </var-tabs-items>